backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
prisma==0.13.1
pydantic==2.9.0
pydantic-settings==2.5.0
python-multipart==0.0.12
google-cloud-aiplatform==1.71.0
aiofiles==24.1.0
python-dotenv==1.0.1
httpx==0.27.0
tenacity==9.0.0
pytest==8.3.4
pytest-asyncio==0.24.0
locust==2.43.2
bcrypt==4.2.0
python-jose[cryptography]==3.3.0